+++ dgisim/src/character/characters_set.py
@@ -13,6 +13,7 @@ _DEFAULT_CHARACTERS: list[type[Character]] = [
     Keqing,
     Kaeya,
     Klee,
+    Mona,
     RhodeiaOfLoch,
     Tighnari,
 ]